Graded HW4
Docstrings are nicely written, but a bit long. Ideally, docstrings are contain what the function does, function input, and function output. No need to write a paragraph. In line comments are good, but a bit much. Try to make sure your code can self-comment via descriptive variable names and good code structure. 
**Documentation: +1**

There are a lot of nested loops. I understand that loops are an easy way to solve problems like this, but see if you can try to format the code so that you can avoid having to do that. **Style: +1**

Code runs as intended. **Functionality: +3**

Overall:
Interesting project, and good job overcoming the hurdles from before. This was an ambitious project, and you pulled it off!

**Grade: 5/5**